+++ frontend/src/api/open-hands.ts
@@ -14,7 +14,7 @@ import {
 } from "./open-hands.types";
 import { openHands } from "./open-hands-axios";
 import { ApiSettings, PostApiSettings } from "#/types/settings";
-import { GitUser, GitRepository } from "#/types/github";
+import { GitUser, GitRepository } from "#/types/git";
 
 class OpenHands {
   /**
