+++ frontend/src/components/features/git/git-repo-selector.tsx
@@ -12,7 +12,7 @@ import { I18nKey } from "#/i18n/declaration";
 import { setSelectedRepository } from "#/state/initial-query-slice";
 import { useConfig } from "#/hooks/query/use-config";
 import { sanitizeQuery } from "#/utils/sanitize-query";
-import { GitRepository } from "#/types/github";
+import { GitRepository } from "#/types/git";
 import { Provider, ProviderOptions } from "#/types/settings";
 
 interface GitRepositorySelectorProps {
